Move ConfigSessionProvider to session project
