+++ core/src/main/resources/reference.conf
@@ -38,7 +38,7 @@ cassandra-journal {
   # ConfigSessionProvider and overriding the lookupContactPoints method.
   # It may optionally have a constructor with an ActorSystem and Config parameter.
   # The config parameter is this config section of the plugin.
-  session-provider = akka.persistence.cassandra.ConfigSessionProvider
+  session-provider = akka.cassandra.session.ConfigSessionProvider
 
   # The identifier that will be passed as parameter to the
   # ConfigSessionProvider.lookupContactPoints method. 
@@ -390,7 +390,7 @@ cassandra-snapshot-store {
   # ConfigSessionProvider and overriding the lookupContactPoints method.
   # It may optionally have a constructor with an ActorSystem and Config parameter.
   # The config parameter is this config section of the plugin.
-  session-provider = akka.persistence.cassandra.ConfigSessionProvider
+  session-provider = akka.cassandra.session.ConfigSessionProvider
 
   # The identifier that will be passed as parameter to the
   # ConfigSessionProvider.lookupContactPoints method.
